Add Java8 template
This adds the Java 8 template for OpenFaaS which uses gradle
4.8.1 as a build system. This has been tested with community
members and validated with Java developers. It mimicks the API
of the other templates but is the first template to be included
which leverages the of-watchdog and HTTP mode.

A README.md is included for future maintenance.

Signed-off-by: Alex Ellis <alexellis2@gmail.com>

request query params support added